pyproject.toml
[project]
name = "expense-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.2",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]
